Migrate config pieces to constants file
